Extracts ConnectionSelection[List]
