+++ react/components/ConnectionSelectionList/components/SelectableChannel/fragments/selectableChannel.js
